Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@888685 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/math/genetics/GeneticAlgorithmTestPermutations.java
@@ -112,7 +112,7 @@ public class GeneticAlgorithmTestPermutations {
             int res = 0;
             List<Integer> decoded = decode(sequence);
             for (int i=0; i<decoded.size(); i++) {
-                int value = (Integer) decoded.get(i);
+                int value = decoded.get(i);
                 if (value != i) {
                     // bad position found
                     res += Math.abs(value - i);
